签到成功活动积分功能

+++ RuoYi-Vue-master/ruoyi-common/src/main/java/com/ruoyi/common/core/domain/entity/SysUser.java
@@ -1,5 +1,6 @@
 package com.ruoyi.common.core.domain.entity;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import javax.validation.constraints.*;
@@ -92,7 +93,7 @@ public class SysUser extends BaseEntity
     /**  地址*/
     private String address;
     /**  积分*/
-    private Double credit;
+    private BigDecimal credit;
 
 
     /** 部门对象 */
@@ -393,11 +394,11 @@ public class SysUser extends BaseEntity
         this.address = address;
     }
 
-    public Double getCredit() {
+    public BigDecimal getCredit() {
         return credit;
     }
 
-    public void setCredit(Double credit) {
+    public void setCredit(BigDecimal credit) {
         this.credit = credit;
     }
